On 15.June.2023
CHanges somethings and added the Courier table Where the member/Admin adds the courier so that he is able to take products delivered from the shops. Now i was on adding the sliding images in the database as i have ended adding the table in the db. The thing i will continue on is to update the delivering method where the seller will first fill the courier's serial numbers then after he will be able to view his image and confirm identity so that he gives him products to deliver.